refactor(video-editor): move transition geometry to models layer
The UI-service boundary ratchet failed because the timeline header and
transition picker imported package:openvine/services/ directly for pure
transition geometry (clampTransitions, renderedOutputDuration,
editorToOutputPosition, transitionConsumedPerSide). These are pure
derivations over DivineVideoClip with no I/O or repository access, so
they belong in the models layer where both the render pipeline and the
UI may import them.

Moves all transition geometry into
lib/models/video_editor/transition_geometry.dart and repoints the render
service, seam render service, canvas, header and picker at it. Relocates
the geometry tests under test/models/ to mirror the new structure.
